--- js/functions.js
@@ -1,6 +1,0 @@
-function checkStringLength(string, maxLength) {
-  return string.length <= maxLength;
-}
-function isPalindrome(str) {
-    return cleanStr === cleanStr.split('').reverse().join('');
-}
